update insertion of expression for affymetrix data

+++ pipeline/Affymetrix/Data_insertion/insert_affy_expression.pl
@@ -2,7 +2,7 @@
 # Julien Roux, created 24/07/09
 # modified 15/08/16, JR: revamped whole script for bgee v15 with condition Ids
 # modified Nov 2016, Frederic Bastian: split former create_files_affy.pl script between insert_affy.pl and insert_affy_expression.pl.
-# last modified Dec. 2016 for use of new microarrayExperimentExpression table
+# last modified March 2021 for bgee 15.0
 
 # Perl core modules
 use strict;
@@ -59,33 +59,63 @@ print "Done, ", scalar(@exprMappedConditions), " conditions retrieved.\n";
 # TO KNOW WHICH PROBESETS ARE NEVER SEEN AS PRESENT #
 #####################################################
 print "Examining all results to update probesets never seen as 'present'...\n";
-my $findPresentGenes = $bgee->prepare('CREATE TEMPORARY TABLE tempPresentAffy (PRIMARY KEY (affymetrixProbesetId, chipTypeId)) ENGINE=InnoDB
-                                        AS (
-                                            SELECT DISTINCT t1.affymetrixProbesetId, t2.chipTypeId
-                                            FROM affymetrixProbeset AS t1
-                                            INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId
-                                            WHERE t1.detectionFlag = "'.$Utils::PRESENT_CALL.'"
-                                        )');
-$findPresentGenes->execute()  or die $findPresentGenes->errstr;
+
+if($debug) { 
+    print "CREATE TEMPORARY TABLE tempPresentAffy (PRIMARY KEY (affymetrixProbesetId, chipTypeId)) ENGINE=InnoDB ",
+          "AS (",
+          "SELECT DISTINCT t1.affymetrixProbesetId, t2.chipTypeId ",
+          "FROM affymetrixProbeset AS t1 ",
+          "INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId ",
+          'WHERE t1.detectionFlag = "'.$Utils::PRESENT_CALL.'")'."\n";
+} else {
+    my $findPresentGenes = $bgee->prepare('CREATE TEMPORARY TABLE tempPresentAffy (PRIMARY KEY (affymetrixProbesetId, chipTypeId)) ENGINE=InnoDB
+                    AS (
+                        SELECT DISTINCT t1.affymetrixProbesetId, t2.chipTypeId
+                        FROM affymetrixProbeset AS t1
+                        INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId
+                        WHERE t1.detectionFlag = "'.$Utils::PRESENT_CALL.'"
+                    )');
+    $findPresentGenes->execute()  or die $findPresentGenes->errstr;
+}
 
 # The "not excluded" status has to be set first, for the next query to properly take into account "undefined" status
-my $presentUp = $bgee->prepare('UPDATE affymetrixProbeset AS t1
+if($debug) { 
+    print 'UPDATE affymetrixProbeset AS t1',
+          'INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId ',
+          'INNER JOIN tempPresentAffy AS t3 ',
+          'ON t1.affymetrixProbesetId = t3.affymetrixProbesetId AND t2.chipTypeId = t3.chipTypeId ',
+          'SET reasonForExclusion = "'.$Utils::CALL_NOT_EXCLUDED.'"'."\n";
+} else {
+    my $presentUp = $bgee->prepare('UPDATE affymetrixProbeset AS t1
                                 INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId
                                 INNER JOIN tempPresentAffy AS t3
                                 ON t1.affymetrixProbesetId = t3.affymetrixProbesetId AND t2.chipTypeId = t3.chipTypeId
                                 SET reasonForExclusion = "'.$Utils::CALL_NOT_EXCLUDED.'"');
-$presentUp->execute()  or die $presentUp->errstr;
-my $preFilteringUp = $bgee->prepare('UPDATE affymetrixProbeset AS t1
+    $presentUp->execute()  or die $presentUp->errstr;
+}
+if($debug) {
+    print 'UPDATE affymetrixProbeset AS t1 ',
+          'INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId ',
+          'LEFT OUTER JOIN tempPresentAffy AS t3 ',
+          'ON t1.affymetrixProbesetId = t3.affymetrixProbesetId AND t2.chipTypeId = t3.chipTypeId ',
+          'SET reasonForExclusion = "'.$Utils::EXCLUDED_FOR_PRE_FILTERED.'" ',
+          'WHERE t3.affymetrixProbesetId IS NULL ',
+          'AND t1.reasonForExclusion != "'.$Utils::EXCLUDED_FOR_UNDEFINED.'"'."\n";
+} else {
+    my $preFilteringUp = $bgee->prepare('UPDATE affymetrixProbeset AS t1
                                      INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId
                                      LEFT OUTER JOIN tempPresentAffy AS t3
                                      ON t1.affymetrixProbesetId = t3.affymetrixProbesetId AND t2.chipTypeId = t3.chipTypeId
                                      SET reasonForExclusion = "'.$Utils::EXCLUDED_FOR_PRE_FILTERED.'"
                                      WHERE t3.affymetrixProbesetId IS NULL AND t1.reasonForExclusion != "'.$Utils::EXCLUDED_FOR_UNDEFINED.'"');
-$preFilteringUp->execute()  or die $preFilteringUp->errstr;
-
-my $dropTempPresent = $bgee->prepare('DROP TABLE tempPresentAffy');
-$dropTempPresent->execute()  or die $dropTempPresent->errstr;
-
+    $preFilteringUp->execute()  or die $preFilteringUp->errstr;
+}
+if ($debug) {
+    print "DROP TABLE tempPresentAffy\n";
+} else {
+    my $dropTempPresent = $bgee->prepare('DROP TABLE tempPresentAffy');
+    $dropTempPresent->execute()  or die $dropTempPresent->errstr;
+}
 print "Done\n";
 
 ##########################################
@@ -104,17 +134,9 @@ my $updResult = $bgee->prepare('UPDATE affymetrixProbeset AS t1
                                 WHERE t1.bgeeGeneId = ? AND t3.exprMappedConditionId = ?
                                     AND t1.reasonForExclusion != "'.$Utils::EXCLUDED_FOR_PRE_FILTERED.'"');
 
-# Insertion into microarrayExperimentExpression
-my $insExpSummary = $bgee->prepare('INSERT INTO microarrayExperimentExpression
-                                    (expressionId, microarrayExperimentId,
-                                        presentHighMicroarrayChipCount, presentLowMicroarrayChipCount,
-                                        absentHighMicroarrayChipCount, absentLowMicroarrayChipCount,
-                                        microarrayExperimentCallDirection, microarrayExperimentCallQuality)
-                                    VALUES (?, ?, ?, ?, ?, ?, ?, ?)');
-
 # query to get all the Affymetrix probesets for a condition
 my $queryResults = $bgee->prepare("SELECT t1.bgeeGeneId, t2.microarrayExperimentId, t1.bgeeAffymetrixChipId,
-                                          t1.affymetrixProbesetId, t1.detectionFlag, t1.affymetrixData
+                                          t1.affymetrixProbesetId, t1.detectionFlag, t1.pValue
                                    FROM affymetrixProbeset AS t1
                                    INNER JOIN affymetrixChip AS t2 ON t1.bgeeAffymetrixChipId = t2.bgeeAffymetrixChipId
                                    INNER JOIN cond AS t3 ON t2.conditionId = t3.conditionId
@@ -145,10 +167,10 @@ for my $exprMappedConditionId ( @exprMappedConditions ){
     $queryResults->execute($exprMappedConditionId)  or die $queryResults->errstr;
     my %results = ();
     while ( my @data = $queryResults->fetchrow_array ){
-        #data[0] = bgeeGeneId, data[1] = experimentId, data[2] = chipId, data[3] = probesetId, data[4] = detectionFlag, data[5] = affymetrixData
+        #data[0] = bgeeGeneId, data[1] = experimentId, data[2] = chipId, data[3] = probesetId, data[4] = detectionFlag, data[5] = pValue
         #important to retrieve the probeset IDs for first per-chip reconciliation
         $results{$data[0]}->{$data[1]}->{$data[2]}->{$data[3]}->{'call'}    = $data[4];
-        $results{$data[0]}->{$data[1]}->{$data[2]}->{$data[3]}->{'quality'} = $data[5];
+        $results{$data[0]}->{$data[1]}->{$data[2]}->{$data[3]}->{'pvalue'} = $data[5];
     }
     print "\t\tDone, ", scalar(keys %results), ' genes retrieved. ',
           "Generating expression summary...\n";
@@ -157,56 +179,8 @@ for my $exprMappedConditionId ( @exprMappedConditions ){
     # (one row for a gene-condition)
     for my $geneId ( keys %results ){
 
-        # aggregate probesets from all chips for this gene/condition to process quality summary
-        my %pbsetsTemp = ();
-        for my $expId ( keys %{$results{$geneId}} ) {
-            my $countTemp  = 0;
-            for my $chip ( keys %{$results{$geneId}->{$expId}} ) {
-                my $pst = 0;
-                my $pst_high_qual = 0;
-                my $abs_high_qual = 0;
-                # First, summarize expression call and quality for all probesets of the gene
-                # pst high > pst low > abs high > abs low
-                for my $pbset ( keys %{$results{$geneId}->{$expId}->{$chip}} ) {
-                    # present wins in all cases: at least one probeset present leads to summarize call as present
-                    if ($results{$geneId}->{$expId}->{$chip}->{$pbset}->{'call'} eq $Utils::PRESENT_CALL ||
-                            $results{$geneId}->{$expId}->{$chip}->{$pbset}->{'call'} eq $Utils::MARGINAL_CALL){
-                        $pst = 1;
-                        # is there at least one present high quality?
-                        if ($results{$geneId}->{$expId}->{$chip}->{$pbset}->{'quality'} eq $Utils::HIGH_QUAL){
-                            $pst_high_qual = 1;
-                        }
-                    }
-                    if ($results{$geneId}->{$expId}->{$chip}->{$pbset}->{'call'} eq $Utils::ABSENT_CALL
-                            and $results{$geneId}->{$expId}->{$chip}->{$pbset}->{'quality'} eq $Utils::HIGH_QUAL){
-                        $abs_high_qual = 1;
-                    }
-                }
-                # add the summarize call by chip to the hash of all calls seen for this gene/experiment/condition
-                if ($pst_high_qual eq 1){
-                    $pbsetsTemp{$expId}->{$countTemp}->{'call'}    = $Utils::PRESENT_CALL;
-                    $pbsetsTemp{$expId}->{$countTemp}->{'quality'} = $Utils::HIGH_QUAL;
-                }
-                elsif ($pst eq 1){
-                    $pbsetsTemp{$expId}->{$countTemp}->{'call'}    = $Utils::PRESENT_CALL;
-                    $pbsetsTemp{$expId}->{$countTemp}->{'quality'} = $Utils::LOW_QUAL;
-                }
-                elsif ($abs_high_qual eq 1){
-                    $pbsetsTemp{$expId}->{$countTemp}->{'call'}    = $Utils::ABSENT_CALL;
-                    $pbsetsTemp{$expId}->{$countTemp}->{'quality'} = $Utils::HIGH_QUAL;
-                }
-                else {
-                    $pbsetsTemp{$expId}->{$countTemp}->{'call'}    = $Utils::ABSENT_CALL;
-                    $pbsetsTemp{$expId}->{$countTemp}->{'quality'} = $Utils::LOW_QUAL;
-                }
-                $countTemp++;
-                #TODO: the above code could be added as a function in insert_expression_utils.pl, receiving as argument the %{$pbsets_summary{$gene}->{'pbsets'}} hash.
-            }
-        }
 
-        # Now, for each gene and condition, we calculate the final quality scores
-        my ($reasonForExclusion, $summary) = Utils::summarizeExperimentCallAndQuality(\%pbsetsTemp);
-
+        my $reasonForExclusion = $Utils::CALL_NOT_EXCLUDED;
         my $expressionId   = undef;
 
         # insert or update the expression table if not only undefined calls
@@ -229,35 +203,9 @@ for my $exprMappedConditionId ( @exprMappedConditions ){
                   "SET expressionId = $printExprId, reasonForExclusion = $reasonForExclusion ",
                   "WHERE t1.bgeeGeneId = $geneId AND t3.exprMappedConditionId = $exprMappedConditionId ",
                   "and t1.reasonForExclusion != '".$Utils::EXCLUDED_FOR_PRE_FILTERED."'\n";
-            if ( $reasonForExclusion eq $Utils::CALL_NOT_EXCLUDED ) {
-                for my $expId ( keys %{ $summary } ) {
-                    print "INSERT INTO microarrayExperimentExpression ",
-                               "(expressionId, microarrayExperimentId, ",
-                               "presentHighMicroarrayChipCount, presentLowMicroarrayChipCount, ",
-                               "absentHighMicroarrayChipCount, absentLowMicroarrayChipCount, ",
-                               "microarrayExperimentCallDirection, microarrayExperimentCallQuality) ",
-                           "VALUES ($printExprId, $expId, ",
-                               "$summary->{$expId}->{'pstHighEvidenceCount'}, ",
-                               "$summary->{$expId}->{'pstLowEvidenceCount'}, ",
-                               "$summary->{$expId}->{'absHighEvidenceCount'}, ",
-                               "$summary->{$expId}->{'absLowEvidenceCount'}, ",
-                               "$summary->{$expId}->{'expCall'}, $summary->{$expId}->{'expCallQuality'})\n";
-                }
-            }
         } else {
             $updResult->execute($expressionId, $reasonForExclusion, $geneId, $exprMappedConditionId)
                 or die $updResult->errstr;
-            if ( $reasonForExclusion eq $Utils::CALL_NOT_EXCLUDED ) {
-                for my $expId ( keys %{ $summary } ) {
-                    $insExpSummary->execute($expressionId, $expId,
-                        $summary->{$expId}->{'pstHighEvidenceCount'},
-                        $summary->{$expId}->{'pstLowEvidenceCount'},
-                        $summary->{$expId}->{'absHighEvidenceCount'},
-                        $summary->{$expId}->{'absLowEvidenceCount'},
-                        $summary->{$expId}->{'expCall'}, $summary->{$expId}->{'expCallQuality'})
-                    or die $insExpSummary->errstr;
-                }
-            }
         }
     }
 }
